Some cleanups
